+++ saleor/graphql/checkout/mutations/__init__.py
@@ -5,6 +5,7 @@ from .checkout_create import CheckoutCreate
 from .checkout_create_from_order import CheckoutCreateFromOrder
 from .checkout_customer_attach import CheckoutCustomerAttach
 from .checkout_customer_detach import CheckoutCustomerDetach
+from .checkout_customer_note_update import CheckoutCustomerNoteUpdate
 from .checkout_delivery_method_update import CheckoutDeliveryMethodUpdate
 from .checkout_email_update import CheckoutEmailUpdate
 from .checkout_language_code_update import CheckoutLanguageCodeUpdate
@@ -25,6 +26,7 @@ __all__ = [
     "CheckoutCreateFromOrder",
     "CheckoutCustomerAttach",
     "CheckoutCustomerDetach",
+    "CheckoutCustomerNoteUpdate",
     "CheckoutDeliveryMethodUpdate",
     "CheckoutEmailUpdate",
     "CheckoutLanguageCodeUpdate",